refactor(display): move interaction logic to Viewport
- Deprecated redundant states (MOVING_PAN, MOVING_TRACKBALL) in favor of a unified AppState::INTERACTION managed dynamically within the update loop.
- Migrated mouse delta and Trackball logic from Scene to Viewport::update using a polling strategy.
- Reduced network overhead by disabling cursor callbacks, preventing event  flooding over the X11 protocol.

+++ 3dEventDisplay/src/scene/ObjectFactory.cpp
@@ -121,7 +121,7 @@ namespace snd3D {
         vertices.push_back(vec3(1.0, 1.0, -1.0));
         vertices.push_back(vec3(-1.0, 1.0, -1.0));
 
-        for (int i = 0; i < vertices.size(); i++) colors.push_back(baseColor);
+        for (int i = 0; i < (int)vertices.size(); i++) colors.push_back(baseColor);
 
         indices.push_back(0); indices.push_back(1); indices.push_back(2);
         indices.push_back(2); indices.push_back(3); indices.push_back(0);
